+++ NEWS.md
@@ -6,6 +6,8 @@ Changes:
 
 Bug Fixes:
 1. Fixed display of error messages in evidence explorer
+2. Fixed searches with backend throwing bad SQL error.
+   Concept sets are now checked to see if all concepts are excluded, with an informative error message.
 
 CemConnector 0.1.0
 ==================